Add Lilya mail sendtest command to use contrib mail
