+++ lilya/cli/constants.py
@@ -2,7 +2,7 @@ LILYA_DISCOVER_APP = "LILYA_DEFAULT_APP"
 DEFAULT_TEMPLATE_NAME = "default"
 APP_PARAMETER = "--app"
 HELP_PARAMETER = "--help"
-EXCLUDED_DIRECTIVES = ["createproject", "createapp", "createdeployment"]
+EXCLUDED_DIRECTIVES = ["createproject", "createapp", "createdeployment", "mail"]
 IGNORE_DIRECTIVES = ["directives"]
 DISCOVERY_FILES = ["application.py", "app.py", "main.py"]
 DISCOVERY_FUNCTIONS = ["get_application", "get_app"]
